eliminate src/lib/ffmpeg.h #339
