+++ src/media/ffmpeg.cpp
@@ -1,8 +1,41 @@
 #include "builddef.h"
 #ifdef USE_FFMPEG
-#include "ffmpeg.h"
+#include <libavutil/error.h>
+#include <libavutil/frame.h>
+#include <libavutil/pixdesc.h>
+#include <libavutil/version.h>
+#include <libavutil/imgutils.h>
+#include <libavutil/rational.h>
+#include <libswscale/swscale.h>
+#include <libswscale/version.h>
+#include <libavformat/version.h>
+#include <libavformat/avformat.h>
 #include "internal.h"
 #include "visual-details.h"
+
+struct AVFormatContext;
+struct AVCodecContext;
+struct AVFrame;
+struct AVCodec;
+struct AVCodecParameters;
+struct AVPacket;
+
+typedef struct ncvisual_details {
+  int packet_outstanding;
+  struct AVFormatContext* fmtctx;
+  struct AVCodecContext* codecctx;     // video codec context
+  struct AVCodecContext* subtcodecctx; // subtitle codec context
+  struct AVFrame* frame;               // frame as read
+  struct AVFrame* oframe;              // RGBA frame
+  struct AVCodec* codec;
+  struct AVCodecParameters* cparams;
+  struct AVCodec* subtcodec;
+  struct AVPacket* packet;
+  struct SwsContext* swsctx;
+  AVSubtitle subtitle;
+  int stream_index;        // match against this following av_read_frame()
+  int sub_stream_index;    // subtitle stream index, can be < 0 if no subtitles
+} ncvisual_details;
 
 #define IMGALLOCALIGN 32
